\name{CD}
\alias{CD}
\docType{data}
\title{Genotype data from a follow-up genetic association study of Crohn's disease}
\description{
This is a real follow-up genetic association data of Crohn's disease. This data contains the genotype frequencies for the initial GWAS scan and the follow-up study.
}
\usage{data(CD);}
\format{
  A data frame containing 12 variables for 11 SNPs.
  \describe{
    \item{\code{DAA1}}{The frequency of homogeneous reference genotype AA in disease group at GWAS scan.}
    \item{\code{DAa1}}{The frequency of heterogeneous genotype Aa in disease group at GWAS scan.}
    \item{\code{Daa1}}{The frequency of homogeneous variant genotype aa in disease group at GWAS scan.}
    \item{\code{NAA1}}{The frequency of homogeneous reference genotype AA in normal group at GWAS scan.}
    \item{\code{NAa1}}{The frequency of heterogeneous genotype Aa in normal group at GWAS scan.}
    \item{\code{Naa1}}{The frequency of homogeneous variant genotype aa in normal group at GWAS scan.}
    \item{\code{DAA2}}{The frequency of homogeneous reference genotype AA in disease group at the follow-up study.}
    \item{\code{DAa2}}{The frequency of heterogeneous genotype Aa in disease group at the follow-up study.}
    \item{\code{Daa2}}{The frequency of homogeneous variant genotype aa in disease group at the follow-up study.}
    \item{\code{NAA2}}{The frequency of homogeneous reference genotype AA in normal group at the follow-up study.}
    \item{\code{NAa2}}{The frequency of heterogeneous genotype Aa in normal group at the follow-up study.}
    \item{\code{Naa2}}{The frequency of homogeneous variant genotype aa in normal group at the follow-up study.}
  }
}
\references{
Parkes M, Barrett JC, Prescott NJ et al. Sequence variants in the autophagy gene IRGM and multiple other replicating loci contribute to Crohn’s disease susceptibility. Nature Genetics 2007; 39(7): 830–832.

Burton PR, Clayton DG, Cardon LR et al. Genome-wide association study of 14,000 cases of seven common diseases and 3,000 shared controls. Nature 2007; 447(7145): 661–678.
}

\keyword{Real data}
\keyword{ significance bias }